Move document to server/
